use more @pretalx/schedule
