+++ webapp/src/main.js
@@ -17,6 +17,7 @@ import 'roboto-fontface'
 import 'roboto-fontface/css/roboto-condensed/roboto-condensed-fontface.css'
 import '@mdi/font/css/materialdesignicons.css'
 import 'quill/dist/quill.core.css'
+import '@pretalx/schedule/dist/schedule.css'
 import 'styles/quill.styl'
 import i18n, { init as i18nInit } from 'i18n'
 import { emojiPlugin } from 'lib/emoji'
